shiny griddy

+++ Field.cpp
@@ -16,8 +16,8 @@ void Field::draw(bool forceRedraw)
 {
 	if (lastDraw != totalTime && !forceRedraw)
 	{
-		int32_t xScreen = (int32_t)(x - world.player.x + SCREEN_WIDTH/2);
-		int32_t yScreen = (int32_t)(y - world.player.y + SCREEN_HEIGHT/2);
+		int32_t xScreen = (int32_t)x - (int32_t)world.player.x + SCREEN_WIDTH/2;
+		int32_t yScreen = (int32_t)y - (int32_t)world.player.y + SCREEN_HEIGHT/2;
 		apply_surface(xScreen, yScreen, image, screen);
 		apply_surface(xScreen, yScreen, image, lightScreen);
